litemap

Module map

Macros

impl_const_get_with_index_for_integer 🔒

litemap_impl 🔒

Structs

KeysIterWrap

An Iterator adapter over the keys of a StoreIterable::KeyValueIter.

LiteMap

A simple “flat” map based on a sorted vector

OccupiedEntry

A view into an occupied entry in a LiteMap.

VacantEntry

A view into a vacant entry in a LiteMap.

ValuesIterWrap

An Iterator adapter over the values of a StoreIterable::KeyValueIter.

Enums

Entry

An entry in a LiteMap, which may be either occupied or vacant.

Functions

const_cmp_bytes 🔒