Trait litemap::store::Store

pub trait Store<K, V> {
    type KeyValueIntoIter: Iterator<Item = (K, V)>;

    fn lm_with_capacity(capacity: usize) -> Self;

    fn lm_reserve(&mut self, additional: usize);

    fn lm_len(&self) -> usize;

    fn lm_get(&self, index: usize) -> Option<(&K, &V)>;

    fn lm_get_mut(&mut self, index: usize) -> Option<(&K, &mut V)>;

    fn lm_binary_search_by<F>(&self, cmp: F) -> Result<usize, usize>
    where
        F: FnMut(&K) -> Ordering;

    fn lm_push(&mut self, key: K, value: V);

    fn lm_insert(&mut self, index: usize, key: K, value: V);

    fn lm_remove(&mut self, index: usize) -> (K, V);

    fn lm_clear(&mut self);

    fn lm_into_iter(self) -> Self::KeyValueIntoIter;

    fn lm_is_empty(&self) -> bool { ... }

    fn lm_last(&self) -> Option<(&K, &V)> { ... }

    fn lm_extend_end(&mut self, other: Self)
    where
        Self: Sized,
    { ... }

    fn lm_extend_start(&mut self, other: Self)
    where
        Self: Sized,
    { ... }

    fn lm_retain<F>(&mut self, predicate: F)
    where
        F: FnMut(&K, &V) -> bool,
    { ... }

}

Trait to enable pluggable backends for LiteMap.

Some methods have default implementations provided for convenience; however, it is generally better to implement all methods that your data store supports.

Associated Types

type KeyValueIntoIter: Iterator<Item = (K, V)>

Required methods

fn lm_with_capacity(capacity: usize) -> Self

Creates a new store with the specified capacity hint.

Implementations may ignore the argument if they do not support pre-allocating capacity.

fn lm_reserve(&mut self, additional: usize)

Reserves additional capacity in the store.

Implementations may ignore the argument if they do not support pre-allocating capacity.

fn lm_len(&self) -> usize

Returns the number of elements in the store.

fn lm_get(&self, index: usize) -> Option<(&K, &V)>

Gets a key/value pair at the specified index.

fn lm_get_mut(&mut self, index: usize) -> Option<(&K, &mut V)>

Gets a key/value pair at the specified index, with a mutable value.

fn lm_binary_search_by<F>(&self, cmp: F) -> Result<usize, usize> where
F: FnMut(&K) -> Ordering,

Searches the store for a particular element with a comparator function.

See the binary search implementation on slice for more information.

fn lm_push(&mut self, key: K, value: V)

Pushes one additional item onto the store.

fn lm_insert(&mut self, index: usize, key: K, value: V)

Inserts an item at a specific index in the store.

Panics

Panics if index is greater than the length.

fn lm_remove(&mut self, index: usize) -> (K, V)

Removes an item at a specific index in the store.

Panics

Panics if index is greater than the length.

fn lm_clear(&mut self)

Removes all items from the store.

fn lm_into_iter(self) -> Self::KeyValueIntoIter

Returns an iterator that moves every item from this store.

Provided methods

fn lm_is_empty(&self) -> bool

Returns whether the store is empty (contains 0 elements).

fn lm_last(&self) -> Option<(&K, &V)>

Gets the last element in the store, or None if the store is empty.

fn lm_extend_end(&mut self, other: Self) where
Self: Sized,

Adds items from another store to the end of this store.

fn lm_extend_start(&mut self, other: Self) where
Self: Sized,

Adds items from another store to the beginning of this store.

fn lm_retain<F>(&mut self, predicate: F) where
F: FnMut(&K, &V) -> bool,

Retains items satisfying a predicate in this store.