litemap::store

Trait StoreIntoIterator

pub trait StoreIntoIterator<K, V>: StoreMut<K, V> {
    type KeyValueIntoIter: Iterator<Item = (K, V)>;

    // Required method
    fn lm_into_iter(self) -> Self::KeyValueIntoIter;

    // Provided methods
    fn lm_extend_end(&mut self, other: Self)
       where Self: Sized { ... }
    fn lm_extend_start(&mut self, other: Self)
       where Self: Sized { ... }
}

Required Associated Types

type KeyValueIntoIter: Iterator<Item = (K, V)>

Required Methods

fn lm_into_iter(self) -> Self::KeyValueIntoIter

Returns an iterator that moves every item from this store.

Provided Methods

fn lm_extend_end(&mut self, other: Self)
where Self: Sized,

Adds items from another store to the end of this store.

fn lm_extend_start(&mut self, other: Self)
where Self: Sized,

Adds items from another store to the beginning of this store.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.