litemap
Module store

Traits for pluggable LiteMap backends.

By default, LiteMap is backed by a Vec. However, in some environments, it may be desirable to use a different data store while still using LiteMap to manage proper ordering of items.

The general guidelines for a performant data store are:

Must support efficient random access for binary search

Should support efficient append operations for deserialization

To plug a custom data store into LiteMap, implement:

Store for most of the methods

StoreIterable for methods that return iterators

StoreFromIterator to enable FromIterator for LiteMap

To test your implementation, enable the "testing" Cargo feature and use check_store().

Traits

Store

Trait to enable pluggable backends for LiteMap.

StoreConstEmpty

Trait to enable const construction of empty store.

StoreFromIterable

StoreFromIterator

A store that can be built from a tuple iterator.

StoreIntoIterator

StoreIterable

Iterator methods for the LiteMap store.

StoreIterableMut

StoreMut

StoreSlice