Trait litemap::store::StoreIterableMut

pub trait StoreIterableMut<'a, K: 'a, V: 'a>: StoreMut<K, V> + StoreIterable<'a, K, V> {
    type KeyValueIterMut: Iterator<Item = (&'a K, &'a mut V)> + DoubleEndedIterator + 'a;
    type KeyValueIntoIter: Iterator<Item = (K, V)>;

    // Required methods
    fn lm_iter_mut(&'a mut self) -> Self::KeyValueIterMut;
    fn lm_into_iter(self) -> Self::KeyValueIntoIter;

    // Provided methods
    fn lm_extend_end(&mut self, other: Self)
       where Self: Sized { ... }
    fn lm_extend_start(&mut self, other: Self)
       where Self: Sized { ... }
}

Required Associated Types

type KeyValueIterMut: Iterator<Item = (&'a K, &'a mut V)> + DoubleEndedIterator + 'a

type KeyValueIntoIter: Iterator<Item = (K, V)>

Required Methods

fn lm_iter_mut(&'a mut self) -> Self::KeyValueIterMut

Returns an iterator over key/value pairs, with a mutable value.

fn lm_into_iter(self) -> Self::KeyValueIntoIter

Returns an iterator that moves every item from this store.

Provided Methods

fn lm_extend_end(&mut self, other: Self)where Self: Sized,

Adds items from another store to the end of this store.

fn lm_extend_start(&mut self, other: Self)where Self: Sized,

Adds items from another store to the beginning of this store.

Object Safety

This trait is not object safe.