Struct litemap::LiteMap

pub struct LiteMap<K, V, S = Vec<(K, V)>> { /* private fields */ }

A simple “flat” map based on a sorted vector

See the module level documentation for why one should use this.

The API is roughly similar to that of std::collections::HashMap, though it requires Ord instead of Hash.

Implementations

impl<K, V> LiteMap<K, V>

pub const fn new() -> Self

Construct a new LiteMap

impl<K, V, S> LiteMap<K, V, S> where
S: Store<K, V>,

pub fn with_capacity(capacity: usize) -> Self

Construct a new LiteMap with a given capacity

pub fn len(&self) -> usize

The number of elements in the LiteMap

pub fn is_empty(&self) -> bool

Whether the LiteMap is empty

pub fn clear(&mut self)

Remove all elements from the LiteMap

pub fn reserve(&mut self, additional: usize)

Reserve capacity for additional more elements to be inserted into the LiteMap to avoid frequent reallocations.

See Vec::reserve() for more information.

pub unsafe fn from_tuple_vec_unchecked(values: S) -> Self

Convert a Vec<(K, V)> into a LiteMap.

Safety

The vec must be sorted and have no duplicate keys.

pub fn into_tuple_vec(self) -> S

Convert a LiteMap into a Vec<(K, V)>.

pub fn get_indexed(&self, index: usize) -> Option<(&K, &V)>

Get the key-value pair residing at a particular index

In most cases, prefer LiteMap::get() over this method.

impl<K, V, S> LiteMap<K, V, S> where
K: Ord,
S: Store<K, V>,

pub fn get<Q: ?Sized>(&self, key: &Q) -> Option<&V> where
K: Borrow<Q>,
Q: Ord,

Get the value associated with key, if it exists.

use litemap::LiteMap;

let mut map = LiteMap::new();
map.insert(1, "one");
map.insert(2, "two");
assert_eq!(map.get(&1), Some(&"one"));
assert_eq!(map.get(&3), None);

pub fn contains_key<Q: ?Sized>(&self, key: &Q) -> bool where
K: Borrow<Q>,
Q: Ord,

Returns whether key is contained in this map

use litemap::LiteMap;

let mut map = LiteMap::new();
map.insert(1, "one");
map.insert(2, "two");
assert_eq!(map.contains_key(&1), true);
assert_eq!(map.contains_key(&3), false);

pub fn get_mut<Q: ?Sized>(&mut self, key: &Q) -> Option<&mut V> where
K: Borrow<Q>,
Q: Ord,

Get the value associated with key, if it exists, as a mutable reference.

use litemap::LiteMap;

let mut map = LiteMap::new();
map.insert(1, "one");
map.insert(2, "two");
if let Some(mut v) = map.get_mut(&1) {
    *v = "uno";
}
assert_eq!(map.get(&1), Some(&"uno"));

pub fn first(&self) -> Option<(&K, &V)>

Get the lowest-rank key/value pair from the LiteMap, if it exists.

Examples

use litemap::LiteMap;

let mut map = LiteMap::new();
assert!(map.try_append(1, "uno").is_none());
assert!(map.try_append(3, "tres").is_none());

assert_eq!(map.first(), Some((&1, &"uno")));

pub fn last(&self) -> Option<(&K, &V)>

Get the highest-rank key/value pair from the LiteMap, if it exists.

Examples

use litemap::LiteMap;

let mut map = LiteMap::new();
assert!(map.try_append(1, "uno").is_none());
assert!(map.try_append(3, "tres").is_none());

assert_eq!(map.last(), Some((&3, &"tres")));

pub fn try_append(&mut self, key: K, value: V) -> Option<(K, V)>

Appends value with key to the end of the underlying vector, returning key and value if it failed. Useful for extending with an existing sorted list.

use litemap::LiteMap;

let mut map = LiteMap::new();
assert!(map.try_append(1, "uno").is_none());
assert!(map.try_append(3, "tres").is_none());

assert!(
    matches!(map.try_append(3, "tres-updated"), Some((3, "tres-updated"))),
    "append duplicate of last key",
);

assert!(
    matches!(map.try_append(2, "dos"), Some((2, "dos"))),
    "append out of order"
);

assert_eq!(map.get(&1), Some(&"uno"));

// contains the original value for the key: 3
assert_eq!(map.get(&3), Some(&"tres"));

// not appended since it wasn't in order
assert_eq!(map.get(&2), None);

pub fn insert(&mut self, key: K, value: V) -> Option<V>

Insert value with key, returning the existing value if it exists.

use litemap::LiteMap;

let mut map = LiteMap::new();
map.insert(1, "one");
map.insert(2, "two");
assert_eq!(map.get(&1), Some(&"one"));
assert_eq!(map.get(&3), None);

pub fn try_insert(&mut self, key: K, value: V) -> Option<(K, V)>

Attempts to insert a unique entry into the map.

If key is not already in the map, inserts it with the corresponding value and returns None.

If key is already in the map, no change is made to the map, and the key and value are returned back to the caller.

use litemap::LiteMap;

let mut map = LiteMap::new();
map.insert(1, "one");
map.insert(3, "three");

// 2 is not yet in the map...
assert_eq!(map.try_insert(2, "two"), None);
assert_eq!(map.len(), 3);

// ...but now it is.
assert_eq!(map.try_insert(2, "TWO"), Some((2, "TWO")));
assert_eq!(map.len(), 3);

pub fn remove<Q: ?Sized>(&mut self, key: &Q) -> Option<V> where
K: Borrow<Q>,
Q: Ord,

Remove the value at key, returning it if it exists.

use litemap::LiteMap;

let mut map = LiteMap::new();
map.insert(1, "one");
map.insert(2, "two");
assert_eq!(map.remove(&1), Some("one"));
assert_eq!(map.get(&1), None);

impl<'a, K: 'a, V: 'a, S> LiteMap<K, V, S> where
K: Ord,
S: StoreIterable<'a, K, V> + StoreFromIterator<K, V>,

pub fn extend_from_litemap(&mut self, other: Self) -> Option<Self>

Insert all elements from other into this LiteMap.

If other contains keys that already exist in self, the values in other replace the corresponding ones in self, and the rejected items from self are returned as a new LiteMap. Otherwise, None is returned.

The implementation of this function is optimized if self and other have no overlap.

Examples

use litemap::LiteMap;

let mut map1 = LiteMap::new();
map1.insert(1, "one");
map1.insert(2, "two");

let mut map2 = LiteMap::new();
map2.insert(2, "TWO");
map2.insert(4, "FOUR");

let leftovers = map1.extend_from_litemap(map2);

assert_eq!(map1.len(), 3);
assert_eq!(map1.get(&1), Some("one").as_ref());
assert_eq!(map1.get(&2), Some("TWO").as_ref());
assert_eq!(map1.get(&4), Some("FOUR").as_ref());

let map3 = leftovers.expect("Duplicate keys");
assert_eq!(map3.len(), 1);
assert_eq!(map3.get(&2), Some("two").as_ref());

impl<K, V, S> LiteMap<K, V, S> where
K: Ord,
S: Store<K, V>,

pub fn find_index<Q: ?Sized>(&self, key: &Q) -> Result<usize, usize> where
K: Borrow<Q>,
Q: Ord,

Obtain the index for a given key, or if the key is not found, the index at which it would be inserted.

(The return value works equivalently to slice::binary_search_by())

The indices returned can be used with Self::get_indexed(). Prefer using Self::get() directly where possible.

impl<'a, K: 'a, V: 'a, S> LiteMap<K, V, S> where
S: StoreIterable<'a, K, V>,

pub fn iter(
&'a self
) -> impl Iterator<Item = (&'a K, &'a V)> + DoubleEndedIterator

Produce an ordered iterator over key-value pairs

pub fn iter_keys(&'a self) -> impl Iterator<Item = &'a K> + DoubleEndedIterator

Produce an ordered iterator over keys

pub fn iter_values(
&'a self
) -> impl Iterator<Item = &'a V> + DoubleEndedIterator

Produce an iterator over values, ordered by their keys

pub fn iter_mut(
&'a mut self
) -> impl Iterator<Item = (&'a K, &'a mut V)> + DoubleEndedIterator

Produce an ordered mutable iterator over key-value pairs

impl<K, V, S> LiteMap<K, V, S> where
S: Store<K, V>,

pub fn retain<F>(&mut self, predicate: F) where
F: FnMut(&K, &V) -> bool,

Retains only the elements specified by the predicate.

In other words, remove all elements such that f((&k, &v)) returns false.

Example

use litemap::LiteMap;

let mut map = LiteMap::new();
map.insert(1, "one");
map.insert(2, "two");
map.insert(3, "three");

// Retain elements with odd keys
map.retain(|k, _| k % 2 == 1);

assert_eq!(map.get(&1), Some(&"one"));
assert_eq!(map.get(&2), None);